+++ line-bot-spring-boot/build.gradle
@@ -17,7 +17,7 @@
 dependencies {
     api project(':line-bot-api-client')
     api project(':line-bot-model')
-    implementation project(':line-bot-servlet')
+    implementation project(':line-bot-parser')
     implementation 'org.springframework.boot:spring-boot-autoconfigure'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'com.google.guava:guava'
